+++ sdk/core/azure-core-test/src/test/java/com/azure/core/test/FakeTestClass.java
@@ -23,6 +23,7 @@ public final class FakeTestClass {
 
     }
 
+    @SuppressWarnings("deprecation")
     @DoNotRecord(skipInPlayback = true)
     public void testWithDoNotRecordSkipInPlayback() {
     }