+++ sdk/openai/azure-ai-openai/src/main/java/com/azure/ai/openai/models/AudioTranscriptionOptions.java
@@ -5,8 +5,8 @@ package com.azure.ai.openai.models;
 
 import com.azure.core.annotation.Fluent;
 import com.azure.core.annotation.Generated;
-import java.util.List;
 import com.azure.core.util.CoreUtils;
+import java.util.List;
 
 /**
  * The configuration information for an audio transcription request.