+++ sdk/openai/azure-ai-openai/src/main/java/com/azure/ai/openai/models/AzureChatExtensionConfiguration.java
@@ -5,31 +5,19 @@ package com.azure.ai.openai.models;
 
 import com.azure.core.annotation.Generated;
 import com.azure.core.annotation.Immutable;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeId;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.fasterxml.jackson.annotation.JsonTypeName;
+import com.azure.json.JsonReader;
+import com.azure.json.JsonSerializable;
+import com.azure.json.JsonToken;
+import com.azure.json.JsonWriter;
+import java.io.IOException;
 
 /**
  * A representation of configuration data for a single Azure OpenAI chat extension. This will be used by a chat
  * completions request that should use Azure OpenAI chat extensions to augment the response behavior.
  * The use of this configuration is compatible only with Azure OpenAI.
  */
-@JsonTypeInfo(
-    use = JsonTypeInfo.Id.NAME,
-    property = "type",
-    defaultImpl = AzureChatExtensionConfiguration.class,
-    visible = true)
-@JsonTypeName("AzureChatExtensionConfiguration")
-@JsonSubTypes({
-    @JsonSubTypes.Type(name = "azure_search", value = AzureSearchChatExtensionConfiguration.class),
-    @JsonSubTypes.Type(name = "azure_ml_index", value = AzureMachineLearningIndexChatExtensionConfiguration.class),
-    @JsonSubTypes.Type(name = "azure_cosmos_db", value = AzureCosmosDBChatExtensionConfiguration.class),
-    @JsonSubTypes.Type(name = "elasticsearch", value = ElasticsearchChatExtensionConfiguration.class),
-    @JsonSubTypes.Type(name = "pinecone", value = PineconeChatExtensionConfiguration.class) })
 @Immutable
-public class AzureChatExtensionConfiguration {
+public class AzureChatExtensionConfiguration implements JsonSerializable<AzureChatExtensionConfiguration> {
 
     /**
      * Creates an instance of AzureChatExtensionConfiguration class.
@@ -44,8 +32,6 @@ public class AzureChatExtensionConfiguration {
      *   Azure chat extensions are only compatible with Azure OpenAI.
      */
     @Generated
-    @JsonTypeId
-    @JsonProperty(value = "type")
     private AzureChatExtensionType type;
 
     /**
@@ -58,5 +44,78 @@ public class AzureChatExtensionConfiguration {
     @Generated
     public AzureChatExtensionType getType() {
         return this.type;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Generated
+    @Override
+    public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
+        jsonWriter.writeStartObject();
+        jsonWriter.writeStringField("type", this.type == null ? null : this.type.toString());
+        return jsonWriter.writeEndObject();
+    }
+
+    /**
+     * Reads an instance of AzureChatExtensionConfiguration from the JsonReader.
+     *
+     * @param jsonReader The JsonReader being read.
+     * @return An instance of AzureChatExtensionConfiguration if the JsonReader was pointing to an instance of it, or
+     * null if it was pointing to JSON null.
+     * @throws IOException If an error occurs while reading the AzureChatExtensionConfiguration.
+     */
+    @Generated
+    public static AzureChatExtensionConfiguration fromJson(JsonReader jsonReader) throws IOException {
+        return jsonReader.readObject(reader -> {
+            String discriminatorValue = null;
+            try (JsonReader readerToUse = reader.bufferObject()) {
+                // Prepare for reading
+                readerToUse.nextToken();
+                while (readerToUse.nextToken() != JsonToken.END_OBJECT) {
+                    String fieldName = readerToUse.getFieldName();
+                    readerToUse.nextToken();
+                    if ("type".equals(fieldName)) {
+                        discriminatorValue = readerToUse.getString();
+                        break;
+                    } else {
+                        readerToUse.skipChildren();
+                    }
+                }
+                // Use the discriminator value to determine which subtype should be deserialized.
+                if ("azure_search".equals(discriminatorValue)) {
+                    return AzureSearchChatExtensionConfiguration.fromJson(readerToUse.reset());
+                } else if ("azure_ml_index".equals(discriminatorValue)) {
+                    return AzureMachineLearningIndexChatExtensionConfiguration.fromJson(readerToUse.reset());
+                } else if ("azure_cosmos_db".equals(discriminatorValue)) {
+                    return AzureCosmosDBChatExtensionConfiguration.fromJson(readerToUse.reset());
+                } else if ("elasticsearch".equals(discriminatorValue)) {
+                    return ElasticsearchChatExtensionConfiguration.fromJson(readerToUse.reset());
+                } else if ("pinecone".equals(discriminatorValue)) {
+                    return PineconeChatExtensionConfiguration.fromJson(readerToUse.reset());
+                } else {
+                    return fromJsonKnownDiscriminator(readerToUse.reset());
+                }
+            }
+        });
+    }
+
+    @Generated
+    static AzureChatExtensionConfiguration fromJsonKnownDiscriminator(JsonReader jsonReader) throws IOException {
+        return jsonReader.readObject(reader -> {
+            AzureChatExtensionConfiguration deserializedAzureChatExtensionConfiguration
+                = new AzureChatExtensionConfiguration();
+            while (reader.nextToken() != JsonToken.END_OBJECT) {
+                String fieldName = reader.getFieldName();
+                reader.nextToken();
+                if ("type".equals(fieldName)) {
+                    deserializedAzureChatExtensionConfiguration.type
+                        = AzureChatExtensionType.fromString(reader.getString());
+                } else {
+                    reader.skipChildren();
+                }
+            }
+            return deserializedAzureChatExtensionConfiguration;
+        });
     }
 }