+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/models/VectorizableTextQuery.java
@@ -77,6 +77,24 @@ public final class VectorizableTextQuery extends VectorQuery {
         return this;
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public VectorizableTextQuery setWeight(Float weight) {
+        super.setWeight(weight);
+        return this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public VectorizableTextQuery setThreshold(VectorThreshold threshold) {
+        super.setThreshold(threshold);
+        return this;
+    }
+
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
@@ -85,6 +103,8 @@ public final class VectorizableTextQuery extends VectorQuery {
         jsonWriter.writeStringField("fields", getFields());
         jsonWriter.writeBooleanField("exhaustive", isExhaustive());
         jsonWriter.writeNumberField("oversampling", getOversampling());
+        jsonWriter.writeNumberField("weight", getWeight());
+        jsonWriter.writeJsonField("threshold", getThreshold());
         jsonWriter.writeStringField("text", this.text);
         return jsonWriter.writeEndObject();
     }
@@ -105,6 +125,8 @@ public final class VectorizableTextQuery extends VectorQuery {
             String fields = null;
             Boolean exhaustive = null;
             Double oversampling = null;
+            Float weight = null;
+            VectorThreshold threshold = null;
             boolean textFound = false;
             String text = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
@@ -125,6 +147,10 @@ public final class VectorizableTextQuery extends VectorQuery {
                     exhaustive = reader.getNullable(JsonReader::getBoolean);
                 } else if ("oversampling".equals(fieldName)) {
                     oversampling = reader.getNullable(JsonReader::getDouble);
+                } else if ("weight".equals(fieldName)) {
+                    weight = reader.getNullable(JsonReader::getFloat);
+                } else if ("threshold".equals(fieldName)) {
+                    threshold = VectorThreshold.fromJson(reader);
                 } else if ("text".equals(fieldName)) {
                     text = reader.getString();
                     textFound = true;
@@ -138,6 +164,8 @@ public final class VectorizableTextQuery extends VectorQuery {
                 deserializedVectorizableTextQuery.setFields(fields);
                 deserializedVectorizableTextQuery.setExhaustive(exhaustive);
                 deserializedVectorizableTextQuery.setOversampling(oversampling);
+                deserializedVectorizableTextQuery.setWeight(weight);
+                deserializedVectorizableTextQuery.setThreshold(threshold);
                 return deserializedVectorizableTextQuery;
             }
             throw new IllegalStateException("Missing required property: text");